pyproject.toml
[tool.poetry]
name = "wool24"
version = "0.1.0-dev"
description = ""
authors = ["Yoga Benggolo <aybenggolo@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
selenium = "^4.9.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.0.270"
black = "^23.3.0"
pylint = "^2.17.4"
mypy = "^1.3.0"
pytest = "^7.3.1"
pylint-pydantic = "^0.1.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
strict = true
warn_unused_ignores = false

plugins = [
    "pydantic.mypy",
]

[tool.ruff]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "C",    # flake8-comprehensions
    "B",    # flake8-bugbear
]
ignore = [
    "E501",     # line too long, handled by black
    "B905",     # kwarg "strict" for zip()
]

[tool.pylint.MASTER]
load-plugins = "pylint_pydantic"

[tool.pylint.BASIC]
good-names = ["f", "n", "i", "ex", "fn", "k", "v", "q"]

[tool.pylint."MESSAGES CONTROL"]
disable = [
    "fixme",
    "logging-fstring-interpolation",
]
extension-pkg-whitelist = "pydantic"
